Fix reflection target class existence validation
We were wrongly assuming that contantizing the class name of the
reflection would be enough to find the target constant of the reflection.
However, it turns out that what AR does is to call `compute_type`
on the original model class, which ends up discovering the reflection
target class by walking out of the namespaces of the original model.

We were not matching that behaviour in our check and thus generating
the wrong RBI file as reported in #603.

Since the code to discover the correct class for the given reflection
target is too complex to replicate in our codebase, I decided to catch
the failure of calling `reflection.klass` and reporting a `NameError`
raised by that call to mean that the target constant doesn't exist.

In order to do this properly, I had to refactor the model validation
code to its own method and group all the checks in there.